use BIGINT everywhere
It's not needed for now, but you never know when it might be

+++ db-migrations/001-initial-db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE git_repositories (
-  id INT NOT NULL PRIMARY KEY,
+  id BIGINT NOT NULL PRIMARY KEY,
   graphql_id TEXT NOT NULL,
   owner TEXT NOT NULL,
   name TEXT NOT NULL,
@@ -11,7 +11,7 @@ COMMENT ON COLUMN git_repositories.id IS 'The databaseId from GitHub';
 CREATE TABLE git_commits (
   id BIGSERIAL NOT NULL PRIMARY KEY,
   graphql_id TEXT NOT NULL,
-  git_repository_id INT NOT NULL REFERENCES git_repositories(id),
+  git_repository_id BIGINT NOT NULL REFERENCES git_repositories(id),
   commit_sha TEXT NOT NULL,
   has_package_manifests BOOLEAN NOT NULL DEFAULT false,
   UNIQUE (git_repository_id, commit_sha)
@@ -27,7 +27,7 @@ CREATE TABLE git_commit_parents (
 CREATE TABLE git_tags (
   id BIGSERIAL NOT NULL PRIMARY KEY,
   graphql_id TEXT NOT NULL,
-  git_repository_id INT NOT NULL REFERENCES git_repositories(id),
+  git_repository_id BIGINT NOT NULL REFERENCES git_repositories(id),
   name TEXT NOT NULL,
   target_git_commit_id BIGINT NOT NULL REFERENCES git_commits(id),
   UNIQUE (git_repository_id, name)
@@ -36,17 +36,17 @@ CREATE TABLE git_tags (
 CREATE TABLE git_branches (
   id BIGSERIAL NOT NULL PRIMARY KEY,
   graphql_id TEXT NOT NULL,
-  git_repository_id INT NOT NULL REFERENCES git_repositories(id),
+  git_repository_id BIGINT NOT NULL REFERENCES git_repositories(id),
   name TEXT NOT NULL,
   target_git_commit_id BIGINT NOT NULL REFERENCES git_commits(id),
   UNIQUE (git_repository_id, name)
 );
 
 CREATE TABLE pull_requests (
-  id INT NOT NULL PRIMARY KEY,
+  id BIGINT NOT NULL PRIMARY KEY,
   graphql_id TEXT NOT NULL,
-  git_repository_id INT NOT NULL REFERENCES git_repositories(id),
-  pr_number INT NOT NULL,
+  git_repository_id BIGINT NOT NULL REFERENCES git_repositories(id),
+  pr_number BIGINT NOT NULL,
   title TEXT NOT NULL,
   is_merged BOOLEAN NOT NULL DEFAULT false,
   is_closed BOOLEAN NOT NULL DEFAULT false,
@@ -59,7 +59,7 @@ COMMENT ON COLUMN pull_requests.id IS 'The databaseId from GitHub';
 
 CREATE TABLE git_commit_pull_requests (
   git_commit_id BIGINT NOT NULL REFERENCES git_commits(id),
-  pull_request_id INT NOT NULL REFERENCES pull_requests(id),
+  pull_request_id BIGINT NOT NULL REFERENCES pull_requests(id),
   PRIMARY KEY (git_commit_id, pull_request_id)
 );
 
@@ -70,7 +70,7 @@ INSERT INTO change_log_entry_kinds (id) VALUES ('breaking'), ('feat'), ('fix'), 
 
 CREATE TABLE change_log_entries (
   id BIGSERIAL NOT NULL PRIMARY KEY,
-  pull_request_id INT NOT NULL REFERENCES pull_requests(id),
+  pull_request_id BIGINT NOT NULL REFERENCES pull_requests(id),
   package_name TEXT NOT NULL,
   sort_order_weight DOUBLE PRECISION NOT NULL,
   kind TEXT NOT NULL REFERENCES change_log_entry_kinds(id),
